Admin bug fixes. Changing user role no longer changes users password and manage users has its own homepage
