--- src/views/AdminHomepage.vue
@@ -1,214 +1,0 @@
-<script setup>
-import { onMounted, ref } from "vue";
-import { useRouter } from "vue-router";
-import UserServices from '../services/UserServices';
-import MenuBar from "../components/MenuBar.vue";
-
-const users = ref(null);
-const isEdit = ref(false);
-const isEditName = ref(false);
-const isEditEmail = ref(false);
-let editedName = ref({
-  firstName: '',
-  lastName: '',
-})
-const newInfo = ref({
-  id: undefined,
-  role: undefined,
-  firstName: undefined,
-  lastName: undefined,
-  email: undefined
-});
-
-const roles = ref({
-  student: 'student',
-  career_serv: 'career services',
-  admin: 'admin'
-})
-
-const snackbar = ref({
-  value: false,
-  color: "",
-  text: "",
-});
-
-
-const router = useRouter();
-
-async function getUsers() {
-  await UserServices.getUser()
-    .then(response => {
-      users.value = response.data;
-    })
-    .catch(error => {
-      console.log(error);
-      snackbar.value.value = true;
-      snackbar.value.color = "error";
-      snackbar.value.text = error.response.data.message;
-    })
-}
-
-async function editRole(selected, item) {
-  // Set the selected user and their new role and update the database
-  newInfo.value = item;
-  newInfo.value.id = item.id;
-  newInfo.value.role = selected;
-  await UserServices.updateUsers(newInfo.value)
-    .then(() => {
-      snackbar.value.value = true;
-      snackbar.value.color = "green";
-      snackbar.value.text = `${newInfo.firstName} updated successfully!`;
-    })
-    .catch((error) => {
-      snackbar.value.value = true;
-      snackbar.value.color = "error";
-      snackbar.value.text = error.response.data.message;
-    });
-  // Refresh the table automatically
-  await getUsers();
-  newInfo.value = ""; // empty newInfo for later use
-}
-
-function editUser(item) {
-  newInfo.value = item;
-  newInfo.value.firstName = item.firstName;
-  newInfo.value.lastName = item.lastName;
-  newInfo.value.email = item.email;
-  newInfo.value.id = item.id;
-  isEditName.value = true;
-}
-
-async function saveUser() {
-  await UserServices.updateUsers(newInfo.value)
-    .then(() => {
-      snackbar.value.value = true;
-      snackbar.value.color = "green";
-      snackbar.value.text = `${newInfo.firstName} updated successfully!`;
-    })
-    .catch((error) => {
-      snackbar.value.value = true;
-      snackbar.value.color = "error";
-      snackbar.value.text = error.response.data.message;
-    });
-  // Refresh the table automatically
-  await getUsers();
-  newInfo.value = ""; // empty newInfo for later use
-  isEditName.value = false;
-}
-
-function closeEdit() {
-  isEditName.value = false;
-}
-</script>
-
-<template>
-  <v-container fill-height fluid>
-    <div id="body">
-      <v-row align="center" class="mb-4" justify="center">
-        <v-col cols="10">
-          <v-card-title class="pl-0 text-h4 font-weight-bold">Admin</v-card-title>
-        </v-col>
-        <v-btn size="x-large" variant="flat">
-          My Dashboard
-        </v-btn>
-      </v-row>
-      <v-row justify="space-around">
-        <v-col cols="8">
-          <div id="resumes">
-            <v-card variant="elevated" flex="no-wrap">
-              <v-row>
-                <v-card-title class="text-h4 font-weight-bold">Resumes</v-card-title>
-              </v-row>
-              <!-- Just dummy text to show where the resumes would be -->
-
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin lobortis ante diam, et aliquam est
-              venenatis eu. Aenean eget sem pretium odio pharetra pulvinar ac a dolor. Aliquam vulputate neque a
-              pellentesque semper. Sed gravida tempus ligula ut varius. Maecenas mollis urna sit amet elit suscipit
-              euismod. Aenean cursus felis lobortis ipsum elementum, ac ultrices risus pellentesque. Vivamus vel
-              fermentum dolor. Maecenas vitae lacus erat. Vestibulum malesuada augue ante, sed scelerisque sapien
-              accumsan ut. Praesent ut lorem efficitur, convallis neque sit amet, gravida lacus. Sed maximus felis sed
-              enim rhoncus vehicula.
-
-              Ut tristique arcu lacus, id rutrum lectus egestas eget. Morbi quam tellus, efficitur id dui eget, pulvinar
-              pharetra turpis. Pellentesque sed risus sit amet augue volutpat imperdiet quis ac lacus. Phasellus
-              faucibus lacus at tincidunt dignissim. Interdum et malesuada fames ac ante ipsum primis in faucibus.
-              Curabitur vel magna eu neque ultrices fermentum facilisis sit amet ex. In accumsan ipsum at nunc blandit
-              placerat. Nulla facilisi. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos
-              himenaeos. Integer mattis accumsan urna, at volutpat nunc.
-            </v-card>
-          </div>
-          <div id="users" style="display: none;">
-            <v-card variant="elevated">
-              <v-row>
-                <v-card-title class="text-h4 font-weight-bold">Users</v-card-title>
-              </v-row>
-              <v-data-table density="compact">
-                <thead>
-                  <tr>
-                    <th>Name
-                    </th>
-                    <th>Role</th>
-                    <th>Email</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <!-- Table for the users and their information -->
-                  <tr v-for="(item) in users" :key="item">
-                    <td>
-                      <v-card-text class="d-inline">{{ item.firstName }} {{ item.lastName }}</v-card-text>
-                    </td>
-                    <td>{{ item.role }}
-                      <!-- Menu to pop up to select which user role to change to  -->
-                      <v-menu transition="scale-transition">
-                        <template template v-slot:activator="{ props }">
-                          <v-btn variant="flat" size="small" icon="mdi-dots-vertical" v-bind="props" />
-                        </template>
-                        <v-list>
-                          <v-list-item v-for="(items, i) in roles" :key="i" :value="i">
-                            <v-list-item-title @click="editRole(i, item)">{{ items }}</v-list-item-title>
-                          </v-list-item>
-                        </v-list>
-                      </v-menu>
-                    </td>
-                    <td>{{ item.email }}</td>
-                    <td>
-                      <v-icon icon="mdi-pencil" size="small" variant="flat" @click="editUser(item)"></v-icon>
-                      <v-icon icon="mdi-delete" />
-                    </td>
-                  </tr>
-                </tbody>
-              </v-data-table>
-              <v-dialog :model-value="isEditName" width="800">
-                <v-card class="rounded-lg elevation-5">
-                  <v-card-item>
-                    <v-card-title class="headline mb-2">
-                      {{ isEditName ? "Edit User" : "" }}
-                    </v-card-title>
-                  </v-card-item>
-                  <v-card-text>
-                    <v-text-field v-model="newInfo.firstName" label="First name" required></v-text-field>
-                    <v-text-field v-model="newInfo.lastName" label="Last name" required></v-text-field>
-                    <v-text-field v-model="newInfo.email" label="Email" required></v-text-field>
-                  </v-card-text>
-                  <v-card-actions>
-                    <v-spacer></v-spacer>
-                    <v-btn variant="flat" color="secondary" @click="isEditName ? closeEdit() : false">Close</v-btn>
-                    <v-btn variant="flat" color="primary" @click="saveUser()">{{ isEditName ? "Update User Profile" :
-                      "" }}</v-btn>
-                  </v-card-actions>
-                </v-card>
-              </v-dialog>
-              <v-row>
-                <v-col cols="auto">
-                  <v-btn @click="getUsers()" color="#000235" variant="flat">
-                    show users
-                  </v-btn>
-                </v-col>
-              </v-row>
-            </v-card>
-          </div>
-        </v-col>
-      </v-row>
-    </div>
-  </v-container>
-</template>
